ICMSLST-1486 Split tests to match module names
The tests for mail.libraries.edifact_validator have been moved to their
own test module, and changed to use the base test class (since the
functions under test are all pure functions, no database setup required).
